Update start.sh
The Problem
The error Could not find a declaration file for module 'playwright-extra-plugin-stealth' means that the "stealth" plugin we installed was written in plain JavaScript. Your project is written in TypeScript, which needs a special "declaration file" (.d.ts) to understand the types of the JavaScript code.

The error message gives us the exact solution: we need to create a new declaration file to tell TypeScript, "Hey, trust me, this module exists."

The Solution
We will add a check to your start.sh script to automatically create this necessary file if it's missing. This makes the entire setup fully automated.